darkfid: Enable consensus task.

+++ src/util/async_util.rs
@@ -2,6 +2,6 @@ use smol::Timer;
 use std::time::Duration;
 
 /// Sleep for any number of seconds.
-pub async fn sleep(seconds: u32) {
+pub async fn sleep(seconds: u64) {
     Timer::after(Duration::from_secs(seconds.into())).await;
 }
